+++ src/ui/pages/capacity/index.js
@@ -1,9 +1,0 @@
-export {
-  capacityManagementPageSystem,
-  CapacityManagementPageSystem
-} from "./CapacityManagementPageSystem.js";
-
-export {
-  capacityManagementView,
-  CapacityManagementView
-} from "./CapacityManagementView.js";