+++ src/ui/pages/channels/index.js
@@ -1,9 +1,0 @@
-export {
-  channelManagementPageSystem,
-  ChannelManagementPageSystem
-} from "./ChannelManagementPageSystem.js";
-
-export {
-  channelManagementView,
-  ChannelManagementView
-} from "./ChannelManagementView.js";