+++ src/ui/pages/command-center/index.js
@@ -1,9 +1,0 @@
-export {
-  operatingCommandCenterPageSystem,
-  OperatingCommandCenterPageSystem
-} from "./OperatingCommandCenterPageSystem.js";
-
-export {
-  operatingCommandCenterView,
-  OperatingCommandCenterView
-} from "./OperatingCommandCenterView.js";