+++ src/ui/pages/customers/index.js
@@ -1,9 +1,0 @@
-export {
-  customerManagementPageSystem,
-  CustomerManagementPageSystem
-} from "./CustomerManagementPageSystem.js";
-
-export {
-  customerManagementView,
-  CustomerManagementView
-} from "./CustomerManagementView.js";