+++ src/ui/pages/employee-promotion/index.js
@@ -1,9 +1,0 @@
-export {
-  employeePromotionPageSystem,
-  EmployeePromotionPageSystem
-} from "./EmployeePromotionPageSystem.js";
-
-export {
-  employeePromotionView,
-  EmployeePromotionView
-} from "./EmployeePromotionView.js";