+++ src/ui/pages/employees/index.js
@@ -1,9 +1,0 @@
-export {
-  employeeManagementPageSystem,
-  EmployeeManagementPageSystem
-} from "./EmployeeManagementPageSystem.js";
-
-export {
-  employeeManagementView,
-  EmployeeManagementView
-} from "./EmployeeManagementView.js";