+++ src/ui/pages/equipment-maintenance/index.js
@@ -1,9 +1,0 @@
-export {
-  equipmentMaintenancePageSystem,
-  EquipmentMaintenancePageSystem
-} from "./EquipmentMaintenancePageSystem.js";
-
-export {
-  equipmentMaintenanceView,
-  EquipmentMaintenanceView
-} from "./EquipmentMaintenanceView.js";