+++ src/ui/pages/equipment/index.js
@@ -1,9 +1,0 @@
-export {
-  equipmentManagementPageSystem,
-  EquipmentManagementPageSystem
-} from "./EquipmentManagementPageSystem.js";
-
-export {
-  equipmentManagementView,
-  EquipmentManagementView
-} from "./EquipmentManagementView.js";