+++ src/ui/pages/finance/index.js
@@ -1,9 +1,0 @@
-export {
-  financeCenterPageSystem,
-  FinanceCenterPageSystem
-} from "./FinanceCenterPageSystem.js";
-
-export {
-  financeCenterView,
-  FinanceCenterView
-} from "./FinanceCenterView.js";