+++ src/ui/pages/honors/index.js
@@ -1,9 +1,0 @@
-export {
-  honorHallPageSystem,
-  HonorHallPageSystem
-} from "./HonorHallPageSystem.js";
-
-export {
-  honorHallView,
-  HonorHallView
-} from "./HonorHallView.js";