+++ src/ui/pages/market-strategy/index.js
@@ -1,9 +1,0 @@
-export {
-  marketStrategyPageSystem,
-  MarketStrategyPageSystem
-} from "./MarketStrategyPageSystem.js";
-
-export {
-  marketStrategyView,
-  MarketStrategyView
-} from "./MarketStrategyView.js";