+++ src/ui/pages/marketing/index.js
@@ -1,9 +1,0 @@
-export {
-  memberMarketingPageSystem,
-  MemberMarketingPageSystem
-} from "./MemberMarketingPageSystem.js";
-
-export {
-  memberMarketingView,
-  MemberMarketingView
-} from "./MemberMarketingView.js";