+++ src/ui/pages/menu-engineering/index.js
@@ -1,9 +1,0 @@
-export {
-  menuEngineeringPageSystem,
-  MenuEngineeringPageSystem
-} from "./MenuEngineeringPageSystem.js";
-
-export {
-  menuEngineeringView,
-  MenuEngineeringView
-} from "./MenuEngineeringView.js";