+++ src/ui/pages/menu-optimization/index.js
@@ -1,9 +1,0 @@
-export {
-  menuOptimizationPageSystem,
-  MenuOptimizationPageSystem
-} from "./MenuOptimizationPageSystem.js";
-
-export {
-  menuOptimizationView,
-  MenuOptimizationView
-} from "./MenuOptimizationView.js";