+++ src/ui/pages/more/index.js
@@ -1,9 +1,0 @@
-export {
-  moreHubPageSystem,
-  MoreHubPageSystem
-} from "./MoreHubPageSystem.js";
-
-export {
-  moreHubView,
-  MoreHubView
-} from "./MoreHubView.js";