+++ src/ui/pages/opening/index.js
@@ -1,9 +1,0 @@
-export {
-  openingSetupPageSystem,
-  OpeningSetupPageSystem
-} from "./OpeningSetupPageSystem.js";
-
-export {
-  openingSetupView,
-  OpeningSetupView
-} from "./OpeningSetupView.js";