+++ src/ui/pages/reputation/index.js
@@ -1,9 +1,0 @@
-export {
-  reputationPageSystem,
-  ReputationPageSystem
-} from "./ReputationPageSystem.js";
-
-export {
-  reputationView,
-  ReputationView
-} from "./ReputationView.js";