+++ src/ui/pages/supply/index.js
@@ -1,9 +1,0 @@
-export {
-  supplyManagementPageSystem,
-  SupplyManagementPageSystem
-} from "./SupplyManagementPageSystem.js";
-
-export {
-  supplyManagementView,
-  SupplyManagementView
-} from "./SupplyManagementView.js";