+++ src/ui/pages/workforce-capacity/index.js
@@ -1,9 +1,0 @@
-export {
-  workforceCapacityPageSystem,
-  WorkforceCapacityPageSystem
-} from "./WorkforceCapacityPageSystem.js";
-
-export {
-  workforceCapacityView,
-  WorkforceCapacityView
-} from "./WorkforceCapacityView.js";